requirements.txt
langchain
streamlit
python-dotenv
openai
st-chat-message
pandas
youtube-transcript-api
beautifulsoup4
tiktoken
